Read data from an ACQ4 HDF5 file and turn it into an R time series

Read data from an ACQ4 HDF5 file and turn it into an R time series

Usage

acq4h52ts(files, name = "data", bit64conversion = "double", channel = 2, ...)

Arguments

files

Input acq4 HDF5 files (often end in .ma or .h5)

name

the name of the HDF5 folder containing the data to read

bit64conversion

What to do with 64 bit integers (this normally comes up with unsigned 32 bit integers for which R has no native support but can be encoded without any loss of precision in R's numeric (8 byte double) type). See h5read() for details.

channel

Which data channel to import

...

Additional arguments passed to h5read()

Value

A ts or mts object